+++ src/olm/account/one_time_keys.rs
@@ -14,6 +14,7 @@
 
 use std::collections::{BTreeMap, HashMap};
 
+use rand::CryptoRng;
 use serde::{Deserialize, Serialize};
 
 use super::PUBLIC_MAX_ONE_TIME_KEYS;
@@ -115,9 +116,12 @@ impl OneTimeKeys {
         (public_key, removed)
     }
 
-    fn generate_one_time_key(&mut self) -> (Curve25519PublicKey, Option<Curve25519PublicKey>) {
+    fn generate_one_time_key<R: CryptoRng>(
+        &mut self,
+        rng: &mut R,
+    ) -> (Curve25519PublicKey, Option<Curve25519PublicKey>) {
         let key_id = KeyId(self.next_key_id);
-        let key = Curve25519SecretKey::new();
+        let key = Curve25519SecretKey::new_with_rng(rng);
         self.insert_secret_key(key_id, key, false)
     }
 
@@ -130,12 +134,16 @@ impl OneTimeKeys {
         !self.unpublished_public_keys.contains_key(key_id)
     }
 
-    pub(super) fn generate(&mut self, count: usize) -> OneTimeKeyGenerationResult {
+    pub(super) fn generate<R: CryptoRng>(
+        &mut self,
+        count: usize,
+        rng: &mut R,
+    ) -> OneTimeKeyGenerationResult {
         let mut removed_keys = Vec::new();
         let mut created_keys = Vec::new();
 
         for _ in 0..count {
-            let (created, removed) = self.generate_one_time_key();
+            let (created, removed) = self.generate_one_time_key(rng);
 
             created_keys.push(created);
             if let Some(removed) = removed {
@@ -195,7 +203,7 @@ mod test {
 
         assert!(store.private_keys.is_empty());
 
-        store.generate(OneTimeKeys::MAX_ONE_TIME_KEYS);
+        store.generate(OneTimeKeys::MAX_ONE_TIME_KEYS, &mut rand::rng());
         assert_eq!(store.unpublished_public_keys.len(), OneTimeKeys::MAX_ONE_TIME_KEYS);
         assert_eq!(store.private_keys.len(), OneTimeKeys::MAX_ONE_TIME_KEYS);
         assert_eq!(store.key_ids_by_key.len(), OneTimeKeys::MAX_ONE_TIME_KEYS);
@@ -216,7 +224,7 @@ mod test {
             store.private_keys.keys().next().copied().expect("Couldn't get the first key ID");
         assert_eq!(oldest_key_id, KeyId(0));
 
-        store.generate(10);
+        store.generate(10, &mut rand::rng());
         assert_eq!(store.unpublished_public_keys.len(), 10);
         assert_eq!(store.private_keys.len(), OneTimeKeys::MAX_ONE_TIME_KEYS);
         assert_eq!(store.key_ids_by_key.len(), OneTimeKeys::MAX_ONE_TIME_KEYS);